+++ docs/CHANGELOG.rst
@@ -24,6 +24,8 @@ Changed
   is added to collection annotation preset
 - When moving entities from one collection to another the annotation fields are
   always added to the new collection annotation preset
+- When copying entities always add annotation fields to the target collection
+  annotation preset
 
 Added
 -----